// AdminApp/src/Models/Step2DTO.ts
interface Step2Dto {
    ownerNameEnglish: string;
    ownerNameMyanmar: string;
    passportPhoto: string;
    ownerNrcNumber: string;
    nrcScan: string;
    nrcScanBack: string;
    passportNumber: string;
    passportScanAttachment: string;
    ownerAddress: string;
    contactEmail: string;
    contactPhoneNumber: string;
    attachment: string;
    companyOrIndividualType: string;
    businessOwnerCitizenType: string;
    permanantAddress: string;
    nrcCode: string,
    nrcPrefix: string,
    nrcType: string,
    nrcNo: string
}
export default Step2Dto;